Cleanup Compiler Warnings
git-svn-id: file:///home/jj/hercules.svn/trunk@6197 956126f8-22a0-4046-8f4a-272fa8102e63

+++ sr.h
@@ -529,7 +529,7 @@ do { \
   if ((_len)) { \
     SR_WRITE_HDR((_file), (_key), (_len)); \
     _rc = SR_WRITE((_buf), 1, (_len), (_file)); \
-    if ((int)_rc != (_len)) SR_WRITE_ERROR; \
+    if ((int)_rc != (int)(_len)) SR_WRITE_ERROR; \
   } else \
     SR_WRITE_HDR((_file), (_key), 0); \
 } while (0)
